mat4: Force/warn Inverted/Transposed result to be checked.
These functions have proven easy to misrepresent usage. They do not
update *this, and should thus be marked const as well as throwing at
least a warning when the result is unused - as this may mistakenly be
seen as the mat4 instance at hand being inverted/Transposed when it is
not.

The aim is to consistently use this macro in more places where intent is
dubious, or where slight mistakes (omitting assignments) are easy to
miss.

As you can see in the diff the macro is already highly successful in
identifying cases where Inverted() is incorrectly used.