Improve FatalError formatting and usage.
Use compile-time concatenated formatting strings to be able to format
one proper message, instead of executing strcat that is not formatted
properly on every occasion.

In CHK_ macros, print the checked statement when it fails. This makes
debugging easier and CHK_ invocations less annoying to read with
explicitly stringified method name (of course, the entire statement is
printed now, not just the name).