CRASHFIX: project_polly.glb causes aligned load on unaligned address!
Vectors are copied to the stack by casting them to a 16-byte aligned
float4 type and using the assignment copy operator. When the buffer and
subsequently the float4s within it are not aligned, SEGFAULTs occur.

Simply replace the load with a memcpy. This function is automatically
inlined to implement the best copy strategy, given the "alignment
guarantees" the source variable makes (or doesn't, in this case).