CUDA: Deduplicate material types by templating the switch-case.
Materials had to be specified twice: Once for the MaterialStore, the
second time for the switch-case. Get rid of this duplication by
defining the available materials (and their enumeration value) only
once, and generate the switch-case and list of types from it.

TODO: I think the idea was to do this _without_ the struct Case, and
instead have a constexpr denoting the type in the struct itself.