RenderSystem: Import altered (compiling!) PBRT parser.
This commit imports the PBRT parser and the necessary definitions around
it. Some conflicting types such as vectors are defined to LH2 types.

All files compile, but the API remains to be implemented. This API is
called for everything defined in a PBRT scene description, and should be
used to fill an LH2 "Host" description to the best of our abilities.